corregida la cunfusion
